Combines options tests and removes unused files
